feat: pull out importer logic, expose JS API
